(REF) Civi/Esm - Use traits
